add option to see time consumption in seconds or milliseconds

+++ Heuristic_selection/src/utils.py
@@ -25,11 +25,11 @@ def filter_file_list_by_max_node(file_list,max_nodes_per_batch):
                 filtered_file_list.append(file)
     return filtered_file_list
 
-def plot_scatter(true_Y,predicted_Y,name="",range=0):
+def plot_scatter(true_Y,predicted_Y,name="",range=0,x_label="True Values",y_label="Predictions"):
     a = plt.axes(aspect='equal')
     plt.scatter(true_Y, predicted_Y)
-    plt.xlabel('True Values')
-    plt.ylabel('Predictions')
+    plt.xlabel(x_label)
+    plt.ylabel(y_label)
     small_lims = [0,range]
     lims = [0, np.max([np.max(true_Y), np.max(predicted_Y)])]
     lims = (lambda : small_lims if range!=0 else lims)()
@@ -60,6 +60,7 @@ def run_eldarica_with_shell_pool_with_file_list(thread,fun,file_list):
 
 def run_eldarica_with_shell(file_and_param):
     if len(file_and_param)!=0:
+        move_file= (lambda : file_and_param[3] if len(file_and_param)>3 else True)()
         file = file_and_param[0]
         eldarica = "../eldarica-graph-generation/eld "
         # file = "../benchmarks/ulimit-test/Problem19_label06_true-unreach-call.c.flat_000.smt2"
@@ -82,7 +83,7 @@ def run_eldarica_with_shell(file_and_param):
         print("extracting " + file_name + " finished")
         # subprocess.call(supplementary_command)
         os.remove(shell_file_name)
-
-        if not os.path.exists(file + ".circles.gv") and os.path.exists(file):
+        #todo: don't move it if it is out of test set
+        if not os.path.exists(file + ".circles.gv") and os.path.exists(file) and move_file==True:
             os.rename(file,"../benchmarks/shell-timemout/"+file_name)
             print("extracting " + file_name + " failed due to time out, move file to shell-timemout")
